+++ app/Http/Controllers/DepenseController.php
@@ -2,15 +2,19 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\ValidableEntityStatus;
 use App\Http\Requests\Depense\DepensePostRequest;
 use App\Http\Requests\Depense\DepensePutRequest;
 use App\Http\Resources\DepenseListResource;
 use App\Http\Resources\DepenseShowResource;
 use App\Http\Resources\DepenseValidationResource;
 use App\Models\Depense;
+use App\Repositories\VisiteRepository;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\DB;
 
 class DepenseController extends Controller
 {
@@ -19,16 +23,67 @@ class DepenseController extends Controller
      */
     public function index(): JsonResource
     {
-        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at', 'status')
-            ->with(['type' => fn(BelongsTo $query) => $query->select('id', 'nom')])->get();
+        $this->authorize('viewAny', Depense::class);
+        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at', 'date_depense', 'status')
+            ->with(['type' => fn(BelongsTo $query) => $query->select('id', 'nom')])
+            ->get();
         return DepenseListResource::collection($depenses);
+    }
+
+    public function stats(): JsonResponse
+    {
+        $this->authorize('viewStats', Depense::class);
+        $depenses = DB::table('depenses')->selectRaw('SUM(montant) as total')
+            ->where('status', ValidableEntityStatus::VALID->value)
+            ->whereBetween('date_depense', [now()->startOfYear(), now()])
+            ->first();
+        $recettes = VisiteRepository::entreesDateFilter([now()->startOfYear(), now()]);
+        return response()->json([
+            'depenses' => [
+                'title' => 'Dépenses',
+                'amount' => (int)$depenses->total,
+            ],
+            'recettes' => [
+                'title' => 'Recettes',
+                'amount' => $recettes,
+            ],
+            'solde' => [
+                'title' => 'Solde',
+                'amount' => $recettes - (int)$depenses->total,
+            ],
+        ]);
     }
 
     public function getPending(): JsonResource
     {
-        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at')
-            ->with(['type' => fn(BelongsTo $query) => $query->select('id', 'nom')])->withResponsible()->pending()->get();
+        $this->authorize('viewPending', Depense::class);
+        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at', 'date_depense')
+            ->with(['type' => fn(BelongsTo $query) => $query->select('id', 'nom')])
+            ->withResponsible()
+            ->pending()
+            ->get();
         return DepenseValidationResource::collection($depenses);
+    }
+
+    public function getPaginate(): JsonResource
+    {
+        $this->authorize('viewAny', Depense::class);
+        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at', 'date_depense', 'status')
+            ->with('type:id,nom')
+            ->latest()
+            ->paginate(8);
+        return DepenseListResource::collection($depenses->withPath('api/depenses/paginate'));
+    }
+
+    public function getSearch(Request $request): JsonResource
+    {
+        $this->authorize('viewAny', Depense::class);
+        $depenses = Depense::select('id', 'titre', 'montant', 'type_depense_id', 'created_at', 'date_depense', 'status')
+            ->with('type:id,nom')
+            ->search($request->search)
+            ->latest()
+            ->paginate(8);
+        return DepenseListResource::collection($depenses->withPath('api/depenses/search'));
     }
 
     /**
@@ -36,6 +91,7 @@ class DepenseController extends Controller
      */
     public function store(DepensePostRequest $request): JsonResponse
     {
+        $this->authorize('create', Depense::class);
         $depense = Depense::make($request->validated());
         $depense->save();
         return response()->json("La dépense $depense->titre a été crée avec succès.");
@@ -46,6 +102,7 @@ class DepenseController extends Controller
      */
     public function show(Depense $depense): JsonResource
     {
+        $this->authorize('view', Depense::class);
         return DepenseShowResource::make($depense->load('type'));
     }
 
@@ -54,6 +111,7 @@ class DepenseController extends Controller
      */
     public function update(DepensePutRequest $request, Depense $depense): JsonResponse
     {
+        $this->authorize('update', Depense::class);
         $depense->update($request->validated());
         return response()->json("La dépense a été modifiée avec succès.");
     }
@@ -63,12 +121,14 @@ class DepenseController extends Controller
      */
     public function destroy(Depense $depense): JsonResponse
     {
+        $this->authorize('delete', Depense::class);
         $depense->delete();
         return response()->json("La dépense $depense->titre a été supprimée avec succès.");
     }
 
     public function valider(Depense $depense): JsonResponse
     {
+        $this->authorize('valider', Depense::class);
         $depense->setValide();
         return response()->json("La depense $depense->titre a été validée avec succès.");
     }
